https://sonarcloud.io/project/issues?id=codewars&issues=AW7-YVvOI_ltG7gakBdn&open=AW7-YVvOI_ltG7gakBdn

+++ kata/7 kyu/easy-line/main/Easyline.java
@@ -4,7 +4,7 @@ class Easyline {
   static BigInteger easyLine(int n) {
     BigInteger c = BigInteger.ONE, sum = BigInteger.ONE;
     for (int i = 0; i <= n; i++) {
-      c = c.multiply(BigInteger.valueOf(n - i)).divide(BigInteger.valueOf(i + 1));
+      c = c.multiply(BigInteger.valueOf(n - (long) i)).divide(BigInteger.valueOf(i + (long)1));
       sum = sum.add(c.multiply(c));
     }
     return sum;
